Выполнены мелкие задачи Спринтов-1-2: (#54)
- проверена работоспособность методов Product;

- инициализированы все List в entity-классах.

- упорядочены аннотации @ToString и @ToString.Exclude

- добавлена проверка использования продукта в заказах перед удалением продукта. Добавлено новое исключение ProductUsedInOrdersException.

- прочие незначительные правки

+++ src/main/java/com/homegarden/store/backend/service/ProductServiceImpl.java
@@ -2,6 +2,7 @@ package com.homegarden.store.backend.service;
 
 import com.homegarden.store.backend.exception.ProductNotFoundException;
 import com.homegarden.store.backend.entity.Product;
+import com.homegarden.store.backend.exception.ProductUsedInOrdersException;
 import com.homegarden.store.backend.repository.ProductRepository;
 import com.homegarden.store.backend.utils.ProductFilterSpecification;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +16,7 @@ import java.util.List;
 public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository productRepository;
+    private final OrderService orderService;
 
     @Override
     public Product create(Product product) {
@@ -43,6 +45,9 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public void delete(Long id) {
         getById(id);
+        if (orderService.isProductUsedInOrders(id)) {
+            throw new ProductUsedInOrdersException("Unable to delete product with id: " + id + " because it is used in Orders");
+        }
         productRepository.deleteById(id);
     }
